add logger

+++ include/dctc/algorithms/connectivity/relays/relays_MST_graph.h
@@ -6,19 +6,19 @@
 
 class RelaysMSTGraph : public MSTGraph {
 protected:
-    double r_c_;
-    double theta_c_;
+    long double r_c_;
+    long double theta_c_;
     int n_total_nodes_omni_;
     std::vector<Node*> terminals_;
     std::vector<Node*> relays_;
-    double beta_;
+    long double beta_;
 
 public:
-    RelaysMSTGraph(const std::vector<Node*>& nodes, GraphNodeType graph_node_type, double r_c, double theta_c,
+    RelaysMSTGraph(const std::vector<Node*>& nodes, GraphNodeType graph_node_type, long double r_c, long double theta_c,
                    std::vector<Edge*>& communication_edges, int n_total_nodes_omni);
     GraphNodeType getGraphNodeType() const;
-    double getThetaC() const;
-    double getRC() const;
+    long double getThetaC() const;
+    long double getRC() const;
     int getNumTerminals() const;
     int getNumRelays() const;
     int getNumTotalNodes() const;
@@ -26,7 +26,7 @@ public:
     std::vector<Node*> getTerminals() const;
     std::vector<Node*> getRelayNodes() const;
     std::vector<Node*> getAllNodes() const;
-    double getBeta() const;
+    long double getBeta() const;
 };
 
 #endif //DCTC_ALGORITHMS_CONNECTIVITY_RELAYS_RELAYS_MST_GRAPH_H_
