+++ include/dctc/instance/instance.h
@@ -28,36 +28,36 @@ private:
     bool regen_targets_;
 
     int n_targets_;
-    double min_range_;
-    double max_range_;
+    long double min_range_;
+    long double max_range_;
     std::vector<Point2D> targets_;
     NodeType node_type_;
-    double r_s_;
-    double r_c_;
-    double theta_s_;
-    double theta_c_;
+    long double r_s_;
+    long double r_c_;
+    long double theta_s_;
+    long double theta_c_;
 
     void init(
-        int n_targets, double min_range, double max_range, 
-        NodeType node_type, double r_s, double r_c, double theta_s, double theta_c,
+        int n_targets, long double min_range, long double max_range, 
+        NodeType node_type, long double r_s, long double r_c, long double theta_s, long double theta_c,
         bool deterministic=false, long unsigned int seed=1, bool regen_targets=true);
     void gen_targets();
 
 public:
     Instance();
     Instance(
-        int n_targets, double min_range, double max_range, 
-        NodeType node_type, double r_s, double r_c, double theta_s, double theta_c,
+        int n_targets, long double min_range, long double max_range, 
+        NodeType node_type, long double r_s, long double r_c, long double theta_s, long double theta_c,
         bool deterministic=false, long unsigned int seed=1, bool regen_targets=true);
     Instance(const Instance& other);
-    double getMinRange() const;
-    double getMaxRange() const;
+    long double getMinRange() const;
+    long double getMaxRange() const;
     bool getDeterministic() const;
     NodeType getNodeType() const;
-    double getRS() const;
-    double getRC() const;
-    double getThetaS() const;
-    double getThetaC() const;
+    long double getRS() const;
+    long double getRC() const;
+    long double getThetaS() const;
+    long double getThetaC() const;
     std::vector<Point2D> getTargets() const;
     void setTargets(const std::vector<Point2D>& targets);
     std::vector<Node*> putCoverageSensors(CoverageAlgType coverage_alg_type) const;
