+++ include/dctc/network_components/nodes/DO_node.h
@@ -10,12 +10,12 @@ class DONode : public Node {
 private:
     Sector sensing_sector_;
     Circle communication_antenna_;
-    void init(double x, double y, NodeType node_type, double r_s, double r_c, double theta_s);
+    void init(long double x, long double y, NodeType node_type, long double r_s, long double r_c, long double theta_s);
     virtual std::string getStringContent() const override;
 public:
     DONode();
-    DONode(double x, double y, NodeType node_type, double r_s, double r_c, double theta_s);
-    DONode(const Point2D& point2D, NodeType node_type, double r_s, double r_c, double theta_s);
+    DONode(long double x, long double y, NodeType node_type, long double r_s, long double r_c, long double theta_s);
+    DONode(const Point2D& point2D, NodeType node_type, long double r_s, long double r_c, long double theta_s);
     DONode(const DONode& other);
     virtual std::string getNodeTypeStr() const override;
     virtual Node* deepCopy(bool preserve_id=false) const override;
