+++ src/dctc/algorithms/connectivity/Aschner/MST_graph_Aschner.cpp
@@ -104,14 +104,14 @@ void MSTGraphAschner::establishCommunicationGraph() {
 
 Node* MSTGraphAschner::findBestNodeToConnect(Node* node, const std::vector<Node*>& candidates) {
     MSTNode* node_to_connect = nullptr;
-    double min_distance = INT_MAX;
+    long double min_distance = INT_MAX;
     Point2D node_position = node->getPoint2D();
     for(Node* candidate : candidates) {
         if (candidate == node) continue;
         if (!nodes_oriented_[candidate]) continue;
         MSTNode* mst_candidate = (MSTNode*) candidate;
         if (!mst_candidate->canCoverOtherNodeByCommunicationAntenna(node)) continue;
-        double distance = computeEuclidDistance(node_position, mst_candidate->getPoint2D());
+        long double distance = computeEuclidDistance(node_position, mst_candidate->getPoint2D());
         if (distance < min_distance) {
             min_distance = distance;
             node_to_connect = mst_candidate;
